Create module for exceptions
