+++ art/attacks/evasion/adversarial_patch.py
@@ -32,7 +32,8 @@ from scipy.ndimage import rotate, shift, zoom
 from art.config import ART_NUMPY_DTYPE
 from art.classifiers.classifier import ClassifierNeuralNetwork, ClassifierGradients
 from art.attacks.attack import EvasionAttack
-from art.utils import check_and_transform_label_format, ClassifierError
+from art.utils import check_and_transform_label_format
+from art.exceptions import ClassifierError
 
 logger = logging.getLogger(__name__)
