+++ art/attacks/evasion/fast_gradient.py
@@ -31,7 +31,7 @@ from art.config import ART_NUMPY_DTYPE
 from art.classifiers.classifier import ClassifierGradients
 from art.attacks.attack import EvasionAttack
 from art.utils import compute_success, get_labels_np_array, random_sphere, projection, check_and_transform_label_format
-from art.utils import ClassifierError
+from art.exceptions import ClassifierError
 
 logger = logging.getLogger(__name__)
