+++ art/attacks/evasion/projected_gradient_descent.py
@@ -33,7 +33,8 @@ from scipy.stats import truncnorm
 from art.config import ART_NUMPY_DTYPE
 from art.classifiers.classifier import ClassifierGradients
 from art.attacks.evasion.fast_gradient import FastGradientMethod
-from art.utils import compute_success, get_labels_np_array, check_and_transform_label_format, ClassifierError
+from art.utils import compute_success, get_labels_np_array, check_and_transform_label_format
+from art.exceptions import ClassifierError
 
 logger = logging.getLogger(__name__)
